fix: cast port to int

+++ config/autoload/server.php
@@ -17,7 +17,7 @@ return [
             'name'      => 'http',
             'type'      => Server::SERVER_HTTP,
             'host'      => '0.0.0.0',
-            'port'      => env('PORT', 9501),
+            'port'      => (int) env('PORT', 9501),
             'sock_type' => SWOOLE_SOCK_TCP,
             'callbacks' => [
                 Event::ON_REQUEST => [\Hyperf\HttpServer\Server::class, 'onRequest'],
